Add files via upload
Project Title: Netflix Homepage Clone

Description:
The Netflix Homepage Clone project for GitHub is a dynamic web application that replicates the iconic Netflix homepage, offering users a realistic and interactive experience akin to the actual streaming platform. This project is designed to showcase front-end development skills, including HTML, CSS, and JavaScript, as well as the integration of external APIs for fetching and displaying content.